Copy in the parent from the category into the object on creation.
The original plan was always to do this, but somewhere along the
line I changed MMObject's get_parent to fall back to getting the
parent from the category.

This adds an extra complication to the inheritance model and
makes it quite difficult - or at least slow - to check that
changes to a category's parent does not create a loop
in the inheritance heirachy. Such a loop is a) pointless, and
b) creates a Stack Exhaustion condition.