Add 'close' method the schema which forces the stores to write all back safely.
The library close all open schema's when __exit__ or it's close is called, and
changes to the FileLogger portion to enssure that all tranasactions are
commited to disk (fsync'd) and that the logfile is removed.

This means that the meas to not consume upwards of 2Gb of diskspace which is
hidden somewhere in /tmp.

Not if you open a packfile and the working directory used by the system to
contain the packfile contents is not deleted on close.

--HG--
extra : source : 32b0bce314fca111b427c45b5c344ca61e2f1652