Change the mercurial interface to use the cmdserver / python-hglib.
Importintg the mercurial internals directly into another program
has never been the recommended way of using mercurial by the hg dev team.
In mercurial 1.9 an cmd server approach was built to provide a stable API, and
python-hglib provides an interfaces to this.

We now use this interface which  means we are not dependent on mercurial
internals the way we used to be.

--HG--
extra : source : 679002d9cde9c21b46894b4ae50f8765eebf4bf8